feat(tabs): simplify graph render
