+++ app/src/pages/Edit.module.css
@@ -11,7 +11,7 @@
 
 .TabsList {
   padding: calc(var(--spacer-px) * 2) calc(var(--spacer-px) * 6) 0;
-  border-bottom: 1px solid var(--color-uiAccent);
+  border-bottom: 2px solid var(--color-lineColor);
 }
 
 .TabsList button {
